+++ lib/akasha/storage/http_event_store/event_serializer.rb
@@ -25,7 +25,7 @@ module Akasha
             metadata = ev['metaData']&.symbolize_keys || {}
             data = ev['data']&.symbolize_keys || {}
             revision = ev['eventNumber']
-            updated_at = Time.parse(ev['updated']) if ev.key?('updated')
+            updated_at = Time.iso8601(ev['updated']) if ev.key?('updated')
             event = Akasha::RecordedEvent.new(ev['eventType'].to_sym, ev['eventId'], revision, updated_at,
                                               metadata, **data)
             event